Unify parse and render into one implementation, update lib.rs to just package the python wrapper and export modules. Move parser/renderer to its own lib file, and have main.rs call this as well resulting in a single implementation between python and rust on the CLI
