+++ src/types/mod.rs
@@ -1,5 +1,7 @@
+mod args;
 mod coords;
 mod star;
 
+pub use args::*;
 pub use coords::*;
 pub use star::*;
